Improve import-time performance
