+++ niveshpy/infrastructure/sqlite/repositories/sqlite_account_repository.py
@@ -10,9 +10,9 @@ from sqlalchemy.orm import Session, sessionmaker
 
 from niveshpy.core.logging import logger
 from niveshpy.core.query.ast import Field, FilterNode
-from niveshpy.core.query.prepare import get_sqlalchemy_filters
 from niveshpy.exceptions import InvalidInputError
 from niveshpy.infrastructure.sqlite.models import Account
+from niveshpy.infrastructure.sqlite.query_filters import get_sqlalchemy_filters
 from niveshpy.models.account import AccountCreate, AccountPublic
