+++ niveshpy/infrastructure/sqlite/repositories/sqlite_security_repository.py
@@ -11,9 +11,9 @@ from sqlalchemy.orm import Session, sessionmaker
 
 from niveshpy.core.logging import logger
 from niveshpy.core.query.ast import Field, FilterNode
-from niveshpy.core.query.prepare import get_sqlalchemy_filters
 from niveshpy.exceptions import ResourceNotFoundError
 from niveshpy.infrastructure.sqlite.models import Security
+from niveshpy.infrastructure.sqlite.query_filters import get_sqlalchemy_filters
 from niveshpy.models.security import SecurityCreate, SecurityPublic
