+++ tests/core/query/test_prepare.py
@@ -4,15 +4,12 @@ from datetime import date
 from decimal import Decimal
 
 import pytest
-from sqlalchemy import column
 
 from niveshpy.core.query.ast import Field, FilterNode, Operator
 from niveshpy.core.query.prepare import (
     combine_filters,
     get_fields_from_queries,
-    get_filters_from_queries,
     group_filters,
-    prepare_expression,
     prepare_filters,
 )
 from niveshpy.exceptions import QuerySyntaxError
@@ -189,295 +186,6 @@ class TestPrepareFilters:
         """Test preparing empty filter list."""
         prepared = prepare_filters([], Field.DESCRIPTION)
         assert len(prepared) == 0
-
-
-class TestPrepareExpression:
-    """Tests for prepare_expression function."""
-
-    @pytest.mark.parametrize(
-        "operator,field,value,col_name",
-        [
-            (Operator.REGEX_MATCH, Field.DESCRIPTION, "test", "description"),
-            (Operator.NOT_REGEX_MATCH, Field.DESCRIPTION, "test", "description"),
-            (Operator.EQUALS, Field.AMOUNT, Decimal("100"), "amount"),
-            (Operator.NOT_EQUALS, Field.AMOUNT, Decimal("100"), "amount"),
-        ],
-        ids=["regex_match", "not_regex_match", "equals", "not_equals"],
-    )
-    def test_binary_operators(self, operator, field, value, col_name):
-        """Test binary operators create appropriate SQLAlchemy expressions."""
-        filter_node = FilterNode(field, operator, value)
-        col = column(col_name)
-
-        expr = prepare_expression(filter_node, col)
-
-        assert hasattr(expr, "type")
-
-    @pytest.mark.parametrize(
-        "operator",
-        [
-            Operator.GREATER_THAN,
-            Operator.GREATER_THAN_EQ,
-            Operator.LESS_THAN,
-            Operator.LESS_THAN_EQ,
-        ],
-    )
-    def test_comparison_operators(self, operator):
-        """Test comparison operators create appropriate expressions."""
-        filter_node = FilterNode(Field.AMOUNT, operator, Decimal("100"))
-        col = column("amount")
-
-        expr = prepare_expression(filter_node, col)
-
-        assert hasattr(expr, "type")
-
-    @pytest.mark.parametrize(
-        "operator,field,value,col_name",
-        [
-            (
-                Operator.BETWEEN,
-                Field.AMOUNT,
-                (Decimal("100"), Decimal("200")),
-                "amount",
-            ),
-            (
-                Operator.NOT_BETWEEN,
-                Field.DATE,
-                (date(2024, 1, 1), date(2024, 12, 31)),
-                "date",
-            ),
-        ],
-        ids=["between", "not_between"],
-    )
-    def test_between_operators(self, operator, field, value, col_name):
-        """Test BETWEEN/NOT_BETWEEN operators create appropriate expressions."""
-        filter_node = FilterNode(field, operator, value)
-        col = column(col_name)
-
-        expr = prepare_expression(filter_node, col)
-
-        assert hasattr(expr, "type")
-
-    @pytest.mark.parametrize(
-        "operator,field,value,col_name",
-        [
-            (
-                Operator.IN,
-                Field.AMOUNT,
-                (Decimal("100"), Decimal("200"), Decimal("300")),
-                "amount",
-            ),
-            (Operator.NOT_IN, Field.TYPE, ("debit", "credit"), "type"),
-        ],
-        ids=["in", "not_in"],
-    )
-    def test_in_operators(self, operator, field, value, col_name):
-        """Test IN/NOT_IN operators create appropriate expressions."""
-        filter_node = FilterNode(field, operator, value)
-        col = column(col_name)
-
-        expr = prepare_expression(filter_node, col)
-
-        assert hasattr(expr, "type")
-
-    def test_unsupported_operator_raises_error(self):
-        """Test unsupported operator/value combination raises OperationError."""
-        from niveshpy.exceptions import OperationError
-
-        # Create invalid filter (e.g., BETWEEN with non-tuple value)
-        filter_node = FilterNode(Field.AMOUNT, Operator.BETWEEN, Decimal("100"))
-        col = column("amount")
-
-        with pytest.raises(OperationError, match="Unsupported operator"):
-            prepare_expression(filter_node, col)
-
-
-class TestGetFiltersFromQueries:
-    """Tests for get_filters_from_queries function."""
-
-    def test_single_query_single_filter(self, column_map):
-        """Test processing single query with one filter."""
-        queries = ("amt:100",)
-
-        expressions = get_filters_from_queries(queries, Field.DESCRIPTION, column_map)
-
-        assert len(expressions) == 1
-
-    def test_multiple_queries(self, column_map):
-        """Test processing multiple query strings."""
-        queries = ("amt:100", "date:2024-01-01")
-
-        expressions = get_filters_from_queries(queries, Field.DESCRIPTION, column_map)
-
-        assert len(expressions) == 2
-
-    def test_default_field_resolution(self, column_map):
-        """Test that DEFAULT field is resolved to default_field parameter."""
-        queries = ("grocery",)
-
-        expressions = get_filters_from_queries(queries, Field.DESCRIPTION, column_map)
-
-        assert len(expressions) == 1
-
-    def test_combining_happens(self, column_map):
-        """Test that multiple EQUALS are combined into IN."""
-        queries = ("amt:100", "amt:200")
-
-        expressions = get_filters_from_queries(queries, Field.DESCRIPTION, column_map)
-
-        # Should be combined into single IN expression
-        assert len(expressions) == 1
-
-    def test_unmapped_field_raises_error(self):
-        """Test that unmapped field raises QuerySyntaxError."""
-        queries = ("sec:AAPL",)
-        # Column map without SECURITY field
-        column_map = {
-            Field.AMOUNT: ["amount"],
-            Field.DATE: ["date"],
-        }
-
-        with pytest.raises(QuerySyntaxError, match="not mapped"):
-            get_filters_from_queries(queries, Field.DESCRIPTION, column_map)
-
-    def test_syntax_error_during_parsing(self):
-        """Test that syntax errors during parsing are caught and re-raised."""
-        queries = ("amt:invalid",)
-        column_map = {Field.AMOUNT: ["amount"]}
-
-        with pytest.raises(QuerySyntaxError):
-            get_filters_from_queries(queries, Field.DESCRIPTION, column_map)
-
-    def test_exception_handler_in_get_filters(self, column_map):
-        """Test that exception handler at lines 177-179 catches and re-raises with context."""
-        queries = ("amt:abc123xyz",)  # Invalid number format
-
-        with pytest.raises(QuerySyntaxError) as exc_info:
-            get_filters_from_queries(queries, Field.DESCRIPTION, column_map)
-
-        # Verify the exception was caught and re-raised
-        # The exception message should contain error information
-        assert "Invalid token" in str(exc_info.value)
-
-    def test_include_fields_none_includes_all(self, column_map):
-        """Test that include_fields=None includes all fields (default behavior)."""
-        queries = ("amt:100", "date:2024-01-01")
-
-        expressions = get_filters_from_queries(
-            queries, Field.DESCRIPTION, column_map, include_fields=None
-        )
-
-        # Both filters should be included
-        assert len(expressions) == 2
-
-    def test_include_fields_includes_matching_fields(self, column_map):
-        """Test that include_fields only includes specified fields."""
-        queries = ("amt:100", "date:2024-01-01", "desc:grocery")
-
-        expressions = get_filters_from_queries(
-            queries, Field.DESCRIPTION, column_map, include_fields={Field.AMOUNT}
-        )
-
-        # Only AMOUNT filter should be included
-        assert len(expressions) == 1
-
-    def test_include_fields_excludes_non_matching_fields(self, column_map):
-        """Test that filters for fields not in include_fields are excluded."""
-        queries = ("amt:100", "date:2024-01-01")
-
-        expressions = get_filters_from_queries(
-            queries,
-            Field.DESCRIPTION,
-            column_map,
-            include_fields={Field.DESCRIPTION, Field.AMOUNT},
-        )
-
-        # DATE filter should be excluded
-        assert len(expressions) == 1
-
-    def test_include_fields_empty_returns_no_expressions(self, column_map):
-        """Test that empty include_fields returns no expressions."""
-        queries = ("amt:100", "date:2024-01-01")
-
-        expressions = get_filters_from_queries(
-            queries, Field.DESCRIPTION, column_map, include_fields=set()
-        )
-
-        # No filters should be included
-        assert len(expressions) == 0
-
-    def test_include_fields_with_multiple_same_field(self, column_map):
-        """Test that include_fields works when multiple filters for same field exist."""
-        queries = ("amt:100", "amt:200", "date:2024-01-01")
-
-        expressions = get_filters_from_queries(
-            queries, Field.DESCRIPTION, column_map, include_fields={Field.AMOUNT}
-        )
-
-        # Combined AMOUNT IN filter should be included, DATE excluded
-        assert len(expressions) == 1
-
-    def test_include_fields_with_default_field(self, column_map):
-        """Test that include_fields works with default field resolution."""
-        queries = ("grocery", "amt:100")
-
-        # DEFAULT field resolves to DESCRIPTION
-        expressions = get_filters_from_queries(
-            queries, Field.DESCRIPTION, column_map, include_fields={Field.DESCRIPTION}
-        )
-
-        # Only DESCRIPTION filter (from "grocery") should be included
-        assert len(expressions) == 1
-
-    def test_text_filters_combined_into_single_expression(self, column_map):
-        """Test that multiple text filters (REGEX_MATCH) are combined into single OR expression."""
-        queries = ("grocery", "restaurant")
-
-        expressions = get_filters_from_queries(queries, Field.DESCRIPTION, column_map)
-
-        # Both text searches should be combined into one OR expression
-        assert len(expressions) == 1
-
-    def test_text_and_non_text_filters_separate(self, column_map):
-        """Test that text filters and non-text filters are separated."""
-        queries = ("grocery", "amt:100")
-
-        expressions = get_filters_from_queries(queries, Field.DESCRIPTION, column_map)
-
-        # Should have 2 expressions: one for text search, one for amount
-        assert len(expressions) == 2
-
-    def test_multiple_text_filters_across_fields(self, column_map):
-        """Test that text filters across different fields are combined."""
-        queries = ("grocery", "acct:savings")
-
-        expressions = get_filters_from_queries(queries, Field.DESCRIPTION, column_map)
-
-        # Text searches (both resolve to REGEX_MATCH) should be combined into one expression
-        # grocery (DESCRIPTION field) and savings (ACCOUNT field)
-        assert len(expressions) == 1
-
-    def test_text_filter_with_multiple_columns(self):
-        """Test that text filter applied to field with multiple columns creates OR'd expressions."""
-        queries = ("test",)
-        # Map DESCRIPTION to multiple columns
-        column_map = {
-            Field.DESCRIPTION: ["desc1", "desc2", "desc3"],
-        }
-
-        expressions = get_filters_from_queries(queries, Field.DESCRIPTION, column_map)
-
-        # All column matches should be OR'd into single expression
-        assert len(expressions) == 1
-
-    def test_mixed_text_and_comparison_operators(self, column_map):
-        """Test mixing text searches with comparison operators."""
-        queries = ("grocery", "amt:>100", "amt:<500")
-
-        expressions = get_filters_from_queries(queries, Field.DESCRIPTION, column_map)
-
-        # Text search (1) + two comparison operators (2) = 3 expressions
-        assert len(expressions) == 3
 
 
 class TestGetFieldsFromQueries:
